+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## next
 
+* Remove double KeyName in SP requests signature
 * Fix Shibboleth SP configuration
 * Hack the metagen.sh script (cloned under /usr/local/bin) in order to match
   the SPID requirements