+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## next
+## 0.4.0 (2018-07-06)
 
 * Fix the order of SP metadata elements
 * Substitute `samlsign` with `xmlsectool`